Add cookie to Ts43AuthToken
Add the cookie from the "Set-Cookie" header of the TS.43 response to
the Ts43AuthToken.
Create new APIs in ServiceEntitlement that return the full HTTP response
for each request instead of just the body as a String.
Modify EapAkaApis to return the full HTTP response instead of just the
body as a String.

Test: atest service-entitlement-tests
Bug: 290658092
Change-Id: If2e7a037ee2643fb686ee39eb2501d3acb3178f4